stockradar/requirements.txt
python-telegram-bot>=20.8
yfinance>=0.2.40
pandas>=2.2
matplotlib>=3.9
mplfinance==0.12.9b7     # 或 0.12.10b0，均為最新 beta
requests>=2.32
twstock>=1.3.3
lxml>=4.9
python-dotenv>=1.0
feedparser>=6.0